console,account: Move human file size util
